feat: Add ping/pong actor example with functional Rust patterns
- Created PingPongExample with PingActor and PongActor
- Added persistent state using rpds::Vector for functional updates
- Implemented zero-panic error handling with Result types
- Added feature-gated examples module in orchestrator
- Added rpds dependency to workspace for persistent data structures
- Implemented functional Rust patterns: map, and_then, no unwraps

Note: Actor trait signatures need alignment with ractor 0.15
- File bead src-3uys for pre-existing reconciler syntax error
- Tests are feature-gated behind 'examples' feature

+++ crates/orchestrator/tests/tier2_random_kill_test.rs
@@ -160,7 +160,7 @@ async fn given_random_tier2_kills_when_100_percent_recovery_required_then_all_re
 
     for cycle in 0..chaos_cycles {
         // Kill a random child
-        let victim_index = (cycle % child_count) as usize;
+        let victim_index = cycle % child_count;
         let victim_name = &child_names[victim_index];
 
         let stop_result = supervisor.cast(SupervisorMessage::StopChild {
